reorganize project as python package
