+++ requirements.txt
@@ -2,3 +2,4 @@ Flask
 flask-wtf
 flask-sqlalchemy
 pymysql
+flask-bcrypt
